First version of MovieRecorder.
This is meant to be used with the Unity video/recorder_with_audio branch.

It introduces an audio input, which is quite simple for now (enable/disable
audio capture). There is disabled functionality in there that we still need
to discuss: it allows capturing the output of arbitrary mixers at the same
time as the main output.

The movie recorder allows to record video and audio into a mp4 or webm
file. Because Unity imports these video files very quickly (no transcoding
by default), the default output path is inside the Assets folder so users
can immediately see and browse their recorded videos using Unity.